requirements.txt
selenium==4.15.2
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
PyPDF2==3.0.1
requests==2.32.3
python-dotenv==1.0.1
webdriver-manager==4.0.2
numpy==2.2.4
pymongo==4.11.3
fastapi==0.116.1
uvicorn==0.35.0
